+++ spec/command/trunk/push_spec.rb
@@ -104,6 +104,7 @@ module Pod
     describe 'validation' do
       before do
         Installer.any_instance.stubs(:aggregate_targets).returns([])
+        Installer.any_instance.stubs(:pod_targets).returns([])
 
         Validator.any_instance.stubs(:check_file_patterns)
         Validator.any_instance.stubs(:validate_url)
@@ -116,6 +117,8 @@ module Pod
         %i(prepare resolve_dependencies download_dependencies).each do |m|
           Installer.any_instance.stubs(m)
         end
+        Command::Trunk::Push.any_instance.stubs(:master_repo_url).
+          returns(Pod::TrunkSource::TRUNK_REPO_URL)
       end
 
       it 'passes the SWIFT_VERSION to the Validator' do
@@ -136,13 +139,13 @@ module Pod
 
       it 'validates specs as frameworks by default' do
         Validator.any_instance.expects(:podfile_from_spec).
-          with(:ios, '8.0', true, []).once.returns(Podfile.new)
+          with(:ios, '8.0', true, [], nil).once.returns(Podfile.new)
         Validator.any_instance.expects(:podfile_from_spec).
-          with(:osx, nil, true, []).once.returns(Podfile.new)
+          with(:osx, nil, true, [], nil).once.returns(Podfile.new)
         Validator.any_instance.expects(:podfile_from_spec).
-          with(:tvos, nil, true, []).once.returns(Podfile.new)
+          with(:tvos, nil, true, [], nil).once.returns(Podfile.new)
         Validator.any_instance.expects(:podfile_from_spec).
-          with(:watchos, nil, true, []).once.returns(Podfile.new)
+          with(:watchos, nil, true, [], nil).once.returns(Podfile.new)
 
         cmd = Command.parse(%w(trunk push spec/fixtures/BananaLib.podspec))
         cmd.send(:validate_podspec)
@@ -150,13 +153,13 @@ module Pod
 
       it 'validates specs as libraries if requested' do
         Validator.any_instance.expects(:podfile_from_spec).
-          with(:ios, nil, false, []).once.returns(Podfile.new)
+          with(:ios, nil, false, [], nil).once.returns(Podfile.new)
         Validator.any_instance.expects(:podfile_from_spec).
-          with(:osx, nil, false, []).once.returns(Podfile.new)
+          with(:osx, nil, false, [], nil).once.returns(Podfile.new)
         Validator.any_instance.expects(:podfile_from_spec).
-          with(:tvos, nil, false, []).once.returns(Podfile.new)
+          with(:tvos, nil, false, [], nil).once.returns(Podfile.new)
         Validator.any_instance.expects(:podfile_from_spec).
-          with(:watchos, nil, false, []).once.returns(Podfile.new)
+          with(:watchos, nil, false, [], nil).once.returns(Podfile.new)
 
         cmd = Command.parse(%w(trunk push spec/fixtures/BananaLib.podspec --use-libraries))
         cmd.send(:validate_podspec)
@@ -231,11 +234,13 @@ module Pod
         @cmd.stubs(:validate_podspec)
         @cmd.stubs(:push_to_trunk).returns([200, success_json])
         Command::Trunk::Push.any_instance.unstub(:update_master_repo)
+        Command::Trunk::Push.any_instance.stubs(:master_repo_name).
+          returns(Pod::TrunkSource::TRUNK_REPO_NAME)
       end
 
       it 'updates the master repo when it exists' do
         Config.instance.sources_manager.stubs(:master_repo_functional?).returns(true)
-        Config.instance.sources_manager.expects(:update).with('master').twice
+        Config.instance.sources_manager.expects(:update).with(Pod::TrunkSource::TRUNK_REPO_NAME).twice
 
         @cmd.run
       end